feat: add availability checking and booking management features

+++ worldcup-hotel-booking/src/main/java/com/worldcup/hotelbooking/booking/booking/BookingMapper.java
@@ -31,7 +31,7 @@ public class BookingMapper {
 
         dto.setCheckInDate(booking.getCheckInDate());
         dto.setCheckOutDate(booking.getCheckOutDate());
-        dto.setTotalPrice(BigDecimal.valueOf(booking.getTotalPrice()));
+        dto.setTotalPrice(booking.getTotalPrice());
         dto.setStatus(booking.getStatus());
         return dto;
     }
